requirements.txt
darts==0.27.0
joblib==1.2.0
matplotlib==3.5.0
numpy==1.20.3
numpy==1.17.4
pandas==1.5.3
pmdarima==2.0.2
python_dateutil==2.8.2
python_dateutil==2.7.3
pytorch_lightning==2.0.8
scikit_learn==1.1.3
seaborn==0.11.2
statsmodels==0.14.0
torch==2.0.1
